worker/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.3.0
nltk==3.8.1